Initial ReadyToRunCodegenNodeFactory cleanup
